+++ src/MiloneCli/ModuleLoad.fs
@@ -35,6 +35,7 @@ open Std.StdSet
 open Std.StdMap
 
 module S = Std.StdString
+module ModuleFetch = MiloneCli.ModuleFetch
 
 type private FetchModuleFun2 =
   ProjectName
@@ -109,8 +110,8 @@ type private RequestResult =
 
 type private RequestMap = TreeMap<ProjectName * ModuleName, RequestResult>
 
-// note: avoid using this function so that DocId can be computed by clients.
-let computeDocId (p: ProjectName) (m: ModuleName) : DocId = Symbol.intern (p + "." + m)
+// #generateDocId
+let computeDocId (p: ProjectName) (m: ModuleName) : DocId = ModuleFetch.computeDocId p m
 
 // -----------------------------------------------
 // ModuleRequest
@@ -246,7 +247,7 @@ let private producer (fetchModule: FetchModuleFun2) (_: State) (r: ModuleRequest
 // Interface
 // -----------------------------------------------
 
-type LoadResult = ModuleSyntaxData2 list list * SyntaxError list
+type LoadResult = ModuleSyntaxData2 list list * DocIdToModulePathMap * SyntaxError list
 
 let load (fetchModule: FetchModuleFun2) (entryProjectName: ProjectName) : LoadResult =
   let entryRequest =
@@ -284,4 +285,16 @@ let load (fetchModule: FetchModuleFun2) (entryProjectName: ProjectName) : LoadRe
 
   let errors = state |> toErrors
 
-  layers, errors
+  let docIdToModulePathMap: DocIdToModulePathMap =
+    layers
+    |> List.fold
+         (fun map layer ->
+           layer
+           |> List.fold
+                (fun map (m: ModuleSyntaxData2) ->
+                  map
+                  |> TMap.add m.DocId (m.ProjectName, m.ModuleName))
+                map)
+         (TMap.empty Symbol.compare)
+
+  layers, docIdToModulePathMap, errors