+++ CHANGELOG.md
@@ -52,6 +52,7 @@
 - Add SyntaxTree generation.
 - Transfer module load process to clients (CLI/LSP server)
 - Revert separation of types and implementation projects
+- `DocId` representation is no longer limited to `P.M`. LSP server now uses absolute file paths to avoid conflict.
 
 ## [v0.5.0] - 2022-07-21